versionCode=19006503  versionName=8.14.0

+++ src/com/yelp/android/services/PersistingCookieStore.java
@@ -2,13 +2,16 @@ package com.yelp.android.services;
 
 import android.text.TextUtils;
 import android.util.Pair;
-import com.crashlytics.android.d;
+import com.yelp.android.util.YelpLog;
 import java.io.File;
 import java.io.FileDescriptor;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
+import java.io.Serializable;
+import java.security.GeneralSecurityException;
 import java.util.Date;
 import javax.crypto.Cipher;
 import javax.crypto.CipherInputStream;
@@ -28,10 +31,11 @@ public class PersistingCookieStore
   {
     b = paramSecretKey;
     a = paramFile;
-    new g(this).start();
+    a();
   }
   
-  ObjectInputStream a(File paramFile)
+  private ObjectInputStream a(File paramFile)
+    throws GeneralSecurityException, IOException
   {
     FileInputStream localFileInputStream = new FileInputStream(paramFile);
     Cipher localCipher = Cipher.getInstance(b.getAlgorithm() + "/CBC/PKCS5Padding");
@@ -39,7 +43,7 @@ public class PersistingCookieStore
     return new ObjectInputStream(new CipherInputStream(localFileInputStream, localCipher));
   }
   
-  public void a()
+  private void a()
   {
     for (;;)
     {
@@ -50,26 +54,29 @@ public class PersistingCookieStore
         {
           Date localDate = new Date();
           File[] arrayOfFile = a.listFiles();
-          int j = arrayOfFile.length;
-          i = 0;
-          if (i < j)
+          if (arrayOfFile != null)
           {
-            File localFile = arrayOfFile[i];
-            try
+            int j = arrayOfFile.length;
+            i = 0;
+            if (i < j)
             {
-              ObjectInputStream localObjectInputStream = a(localFile);
-              Cookie localCookie = (Cookie)localObjectInputStream.readObject();
-              localObjectInputStream.close();
-              if (localCookie.isExpired(localDate)) {
-                localFile.delete();
-              } else {
-                super.addCookie(localCookie);
+              File localFile = arrayOfFile[i];
+              try
+              {
+                ObjectInputStream localObjectInputStream = a(localFile);
+                Cookie localCookie = (Cookie)localObjectInputStream.readObject();
+                localObjectInputStream.close();
+                if (localCookie.isExpired(localDate)) {
+                  localFile.delete();
+                } else {
+                  super.addCookie(localCookie);
+                }
               }
-            }
-            catch (Exception localException)
-            {
-              d.a(localException);
-              localFile.delete();
+              catch (Exception localException)
+              {
+                YelpLog.remoteError(localException);
+                localFile.delete();
+              }
             }
           }
         }
@@ -80,7 +87,7 @@ public class PersistingCookieStore
     }
   }
   
-  void a(Cookie paramCookie)
+  private void a(Cookie paramCookie)
   {
     if (TextUtils.isEmpty(paramCookie.getValue()))
     {
@@ -90,7 +97,7 @@ public class PersistingCookieStore
     try
     {
       Pair localPair = c(paramCookie);
-      ((ObjectOutputStream)first).writeObject(new PersistingCookieStore.SerializableCookie(paramCookie));
+      ((ObjectOutputStream)first).writeObject(new SerializableCookie(paramCookie));
       ((ObjectOutputStream)first).flush();
       ((FileDescriptor)second).sync();
       ((ObjectOutputStream)first).close();
@@ -98,9 +105,36 @@ public class PersistingCookieStore
     }
     catch (Exception localException)
     {
-      d.a(localException);
+      YelpLog.remoteError(localException);
       b(paramCookie);
     }
+  }
+  
+  private IvParameterSpec b(File paramFile)
+  {
+    byte[] arrayOfByte = new byte[16];
+    paramFile = paramFile.getName().getBytes();
+    System.arraycopy(paramFile, 0, arrayOfByte, 0, Math.min(paramFile.length, arrayOfByte.length));
+    return new IvParameterSpec(arrayOfByte);
+  }
+  
+  private void b(Cookie paramCookie)
+  {
+    paramCookie = new File(a, paramCookie.getName());
+    if (!paramCookie.delete()) {
+      paramCookie.deleteOnExit();
+    }
+  }
+  
+  private Pair<ObjectOutputStream, FileDescriptor> c(Cookie paramCookie)
+    throws IOException, GeneralSecurityException
+  {
+    a.mkdirs();
+    paramCookie = new File(a, paramCookie.getName());
+    FileOutputStream localFileOutputStream = new FileOutputStream(paramCookie);
+    Cipher localCipher = Cipher.getInstance(b.getAlgorithm() + "/CBC/PKCS5Padding");
+    localCipher.init(1, b, b(paramCookie));
+    return Pair.create(new ObjectOutputStream(new CipherOutputStream(localFileOutputStream, localCipher)), localFileOutputStream.getFD());
   }
   
   public void addCookie(Cookie paramCookie)
@@ -143,32 +177,6 @@ public class PersistingCookieStore
     }
   }
   
-  IvParameterSpec b(File paramFile)
-  {
-    byte[] arrayOfByte = new byte[16];
-    paramFile = paramFile.getName().getBytes();
-    System.arraycopy(paramFile, 0, arrayOfByte, 0, Math.min(paramFile.length, arrayOfByte.length));
-    return new IvParameterSpec(arrayOfByte);
-  }
-  
-  void b(Cookie paramCookie)
-  {
-    paramCookie = new File(a, paramCookie.getName());
-    if (!paramCookie.delete()) {
-      paramCookie.deleteOnExit();
-    }
-  }
-  
-  Pair<ObjectOutputStream, FileDescriptor> c(Cookie paramCookie)
-  {
-    a.mkdirs();
-    paramCookie = new File(a, paramCookie.getName());
-    FileOutputStream localFileOutputStream = new FileOutputStream(paramCookie);
-    Cipher localCipher = Cipher.getInstance(b.getAlgorithm() + "/CBC/PKCS5Padding");
-    localCipher.init(1, b, b(paramCookie));
-    return Pair.create(new ObjectOutputStream(new CipherOutputStream(localFileOutputStream, localCipher)), localFileOutputStream.getFD());
-  }
-  
   public void clear()
   {
     for (;;)
@@ -200,6 +208,103 @@ public class PersistingCookieStore
       finally {}
       label84:
       i += 1;
+    }
+  }
+  
+  private static class SerializableCookie
+    implements Serializable, Cookie
+  {
+    private static final long serialVersionUID = 1L;
+    private String mComment;
+    private String mCommentUrl;
+    private String mDomain;
+    private Date mExpiryDate;
+    private String mName;
+    private String mPath;
+    private int[] mPorts;
+    private boolean mSecure;
+    private String mValue;
+    private int mVersion;
+    
+    public SerializableCookie() {}
+    
+    public SerializableCookie(String paramString1, String paramString2, int[] paramArrayOfInt, String paramString3, boolean paramBoolean, String paramString4, Date paramDate, String paramString5, int paramInt, String paramString6)
+    {
+      mName = paramString1;
+      mComment = paramString2;
+      mPorts = paramArrayOfInt;
+      mDomain = paramString3;
+      mSecure = paramBoolean;
+      mPath = paramString4;
+      mExpiryDate = paramDate;
+      mValue = paramString5;
+      mVersion = paramInt;
+      mCommentUrl = paramString6;
+    }
+    
+    public SerializableCookie(Cookie paramCookie)
+    {
+      this(paramCookie.getName(), paramCookie.getComment(), paramCookie.getPorts(), paramCookie.getDomain(), paramCookie.isSecure(), paramCookie.getPath(), paramCookie.getExpiryDate(), paramCookie.getValue(), paramCookie.getVersion(), paramCookie.getCommentURL());
+    }
+    
+    public String getComment()
+    {
+      return mComment;
+    }
+    
+    public String getCommentURL()
+    {
+      return mCommentUrl;
+    }
+    
+    public String getDomain()
+    {
+      return mDomain;
+    }
+    
+    public Date getExpiryDate()
+    {
+      return mExpiryDate;
+    }
+    
+    public String getName()
+    {
+      return mName;
+    }
+    
+    public String getPath()
+    {
+      return mPath;
+    }
+    
+    public int[] getPorts()
+    {
+      return mPorts;
+    }
+    
+    public String getValue()
+    {
+      return mValue;
+    }
+    
+    public int getVersion()
+    {
+      return mVersion;
+    }
+    
+    public boolean isExpired(Date paramDate)
+    {
+      return paramDate.after(mExpiryDate);
+    }
+    
+    public boolean isPersistent()
+    {
+      return true;
+    }
+    
+    public boolean isSecure()
+    {
+      return mSecure;
     }
   }
 }
